requirements.txt
streamlit
PyMuPDF
opencv-python-headless
numpy
Pillow